fix: numeric data types

+++ domain/recipe.go
@@ -9,8 +9,8 @@ type RecipeDetails struct {
 	Name        string
 	Description string
 	CreatedBy   *User
-	Servings    int32
-	Minutes     int32
+	Servings    int64
+	Minutes     int64
 }
 
 type StepIngredient struct {
@@ -20,13 +20,13 @@ type StepIngredient struct {
 }
 
 type RecipeStep struct {
-	ID           int32
+	ID           int64
 	Instructions string
 	Ingredients  []StepIngredient
 }
 
 type Recipe struct {
-	ID        int32
+	ID        int64
 	CreatedAt time.Time
 	Tags      []Tag
 	Images    []RecipeImage
@@ -35,7 +35,7 @@ type Recipe struct {
 }
 
 type RecipeImage struct {
-	ID  int32
+	ID  int64
 	URL *url.URL
 }
 
@@ -45,41 +45,41 @@ type MealPlan struct {
 }
 
 type MealPlanRecipe struct {
-	ID     int32
+	ID     int64
 	Recipe Recipe
 }
 
 type MealPlanEntry struct {
-	UserID    int32
-	RecipeID  int32
+	UserID    int64
+	RecipeID  int64
 	Date      time.Time
-	SortOrder int32
+	SortOrder int64
 }
 
 type Ingredient struct {
-	ID        int32
+	ID        int64
 	Name      string
 	Nutrients []IngredientNutrient
 }
 
 type Nutrient struct {
-	ID   int32
+	ID   int64
 	Name string
 	Unit string
 }
 
 type IngredientNutrient struct {
 	Nutrient Nutrient
-	Amount   float32
+	Amount   float64
 }
 
 type Unit struct {
-	ID     int32
+	ID     int64
 	Name   string
 	Symbol *string
 }
 
 type Tag struct {
-	ID   int32
+	ID   int64
 	Name string
 }
